refactor: extract MarketMindAIPage chat/artifact components (F5)
Extract the six presentational components (SectionLabel,
StarterPromptButton, MessageBubble, ContextCard, EvidencePanel,
ArtifactPreview) and their two component-only helpers (ARTIFACT_SECTIONS,
normalizeAssistantContent) into components/marketmind-ai/components.js.
The page keeps its chat state/streaming logic and main-only helpers
(formatActionError, getSentimentToneWord, noticeToneClassNames), dropping
to 775 lines. Pure code movement, guarded by the existing 5-test suite;
lint + build clean.
